refactor(audit): extract tests to log_tests.rs, add changelog
Moves ~460 lines of tests from log.rs (1049 lines) to a sibling
log_tests.rs file via #[path] attribute. log.rs is now 590 lines.

Adds changelog entry for per-principal audit chain splitting.
